[SPARK-47347][PYTHON][CONNECT][TESTS] Factor session-related tests out of `test_connect_basic`
### What changes were proposed in this pull request?
Factor session-related tests out of `test_connect_basic`

### Why are the changes needed?
for testing parallelism

### Does this PR introduce _any_ user-facing change?
no, test-only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #45472 from zhengruifeng/add_test_connect_session.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
